+++ packages/core/src/typings/transaction.ts
@@ -1,5 +1,6 @@
 /**
  * Original work Copyright (c) 2019 Burst Apps Team
+ * Modified work Copyright (c) 2022 Signum Network
  */
 
 /**
@@ -8,30 +9,31 @@
  * @module core
  * */
 export interface Transaction {
-    readonly senderPublicKey?: string;
-    readonly signature?: string;
-    readonly feeNQT?: string;
-    readonly requestProcessingTime?: number;
-    readonly type?: number;
-    readonly confirmations?: number;
-    readonly fullHash?: string;
-    readonly version?: number;
-    readonly ecBlockId?: string;
-    readonly signatureHash?: string;
+    readonly amountNQT: string;
     readonly attachment?: any;
-    readonly senderRS?: string;
-    readonly subtype?: number;
-    readonly amountNQT?: string;
-    readonly sender?: string;
-    readonly ecBlockHeight?: number;
-    readonly block?: string;
-    readonly blockTimestamp?: number;
-    readonly deadline?: number;
-    readonly transaction?: string;
-    readonly timestamp?: number;
-    readonly height?: number;
+    readonly block: string;
+    readonly blockTimestamp: number;
+    readonly cashbackId: string;
+    readonly confirmations?: number;
+    readonly deadline: number;
+    readonly ecBlockHeight: number;
+    readonly ecBlockId: string;
+    readonly feeNQT: string;
+    readonly fullHash: string;
+    readonly height: number;
     readonly recipient?: string;
     readonly recipientRS?: string;
     readonly referencedTransactionFullHash?: string;
+    readonly requestProcessingTime: number;
+    readonly sender: string;
+    readonly senderPublicKey: string;
+    readonly senderRS: string;
+    readonly signature: string;
+    readonly signatureHash: string;
+    readonly subtype: number;
+    readonly timestamp: number;
+    readonly transaction: string;
+    readonly type: number;
+    readonly version: number;
 }